+++ actingweb/aw_proxy.py
@@ -485,7 +485,9 @@ class AwProxy:
             self.last_location = response.headers["Location"]
         else:
             self.last_location = None
-        logger.debug(f"Create trust peer resource async response: {response.status_code}")
+        logger.debug(
+            f"Create trust peer resource async response: {response.status_code}"
+        )
         if response.status_code < 200 or response.status_code > 299:
             logger.warning("Not able to create new trust peer resource async.")
         try:
@@ -577,7 +579,9 @@ class AwProxy:
                     "message": "Internal error communicating with trust peer service.",
                 },
             }
-        logger.debug(f"Change trust peer resource async response: {response.status_code}")
+        logger.debug(
+            f"Change trust peer resource async response: {response.status_code}"
+        )
         if response.status_code < 200 or response.status_code > 299:
             logger.warning("Not able to change trust peer resource async.")
         try:
@@ -654,7 +658,9 @@ class AwProxy:
                     "message": "Internal error communicating with trust peer service.",
                 },
             }
-        logger.debug(f"Delete trust peer resource async response: {response.status_code}")
+        logger.debug(
+            f"Delete trust peer resource async response: {response.status_code}"
+        )
         if response.status_code < 200 or response.status_code > 299:
             logger.warning("Not able to delete trust peer resource async.")
         try: